+++ server/alerter.mjs
@@ -9,6 +9,7 @@ export const processAlerter = async (config) => {
         telegramToken,
         telegramChatIDAlert,
         blockDiff,
+        blockDiffToRestart,
         restartAfterMax,
         restartAfterUnv,
         restartAfterPrev,
@@ -79,52 +80,40 @@ export const processAlerter = async (config) => {
             let message
 
             if (DIFF_MAX >= blockDiff) {
-                OK_MAX = false
                 message = `Difference block height detected!\nHeight ${DIFF_MAX > 0 ? 'less' : 'more'} than max block length!\nDifference: ${Math.abs(DIFF_MAX)}\nNode: ${nHeight}\nMax: ${mHeight} ${sign}`
                 await telegram(message, {token: telegramToken, recipients: telegramChatIDAlert})
 
                 if (restartStateSyncedRules.includes("MAX")) {
-                    if (globalThis.restartTimerMax / 60000 >= restartAfterMax) {
-                        globalThis.restartTimerMax = 0
+                    if (DIFF_MAX >= blockDiffToRestart) {
                         if (canRestartNode && restartCmd) {
                             restart('Difference to max block length!')
                         }
-                    } else {
-                        globalThis.restartTimerMax += alertInterval
                     }
                 }
             }
 
             if (DIFF_UNVALIDATED >= blockDiff) {
-                OK_UNV = false
                 message = `Fork detected!\nHeight ${DIFF_UNVALIDATED > 0 ? 'less' : 'more'} than unvalidated block length!\nDifference: ${Math.abs(DIFF_UNVALIDATED)}\nNode: ${nHeight}\nUnvalidated: ${uHeight} ${sign}`
                 await telegram(message, {token: telegramToken, recipients: telegramChatIDAlert})
 
-                if (restartStateSyncedRules.includes("UNV")) {
-                    if (globalThis.restartTimerUnv / 60000 >= restartAfterUnv) {
-                        globalThis.restartTimerUnv = 0
+                if (restartStateSyncedRules.includes("UNV") || restartStateSyncedRules.includes("FORK")) {
+                    if (DIFF_UNVALIDATED >= blockDiffToRestart) {
                         if (canRestartNode && restartCmd) {
                             restart('Node in fork!')
                         }
-                    } else {
-                        globalThis.restartTimerUnv += alertInterval
                     }
                 }
             }
 
             if (uHeight && DIFF_UNVALIDATED < 0 && Math.abs(DIFF_UNVALIDATED) >= blockDiff) {
-                OK_UNV = false
                 message = `Forward fork detected!\nHeight ${DIFF_UNVALIDATED > 0 ? 'less' : 'more'} than unvalidated block length!\nDifference: ${Math.abs(DIFF_UNVALIDATED)}\nNode: ${nHeight}\nUnvalidated: ${uHeight} ${sign}`
                 await telegram(message, {token: telegramToken, recipients: telegramChatIDAlert})
 
-                if (restartStateSyncedRules.includes("UNV")) {
-                    if (globalThis.restartTimerUnv / 60000 >= restartAfterUnv) {
-                        globalThis.restartTimerUnv = 0
+                if (restartStateSyncedRules.includes("UNV") || restartStateSyncedRules.includes("FORWARD-FORK")) {
+                    if (Math.abs(DIFF_UNVALIDATED) >= blockDiffToRestart) {
                         if (canRestartNode && restartCmd) {
                             restart('Node in forward fork!')
                         }
-                    } else {
-                        globalThis.restartTimerUnv += alertInterval
                     }
                 }
             }
@@ -135,7 +124,7 @@ export const processAlerter = async (config) => {
                     message = `Hanging node detected!\nBlock height ${nHeight} equal to previous value! ${sign}`
                     await telegram(message, {token: telegramToken, recipients: telegramChatIDAlert})
 
-                    if (restartStateSyncedRules.includes("PREV")) {
+                    if (restartStateSyncedRules.includes("PREV") || restartStateSyncedRules.includes("HANG")) {
                         if (globalThis.restartTimerPrev >= restartAfterPrev) {
                             globalThis.restartTimerPrev = 0
                             if (canRestartNode && restartCmd) {
@@ -162,8 +151,6 @@ export const processAlerter = async (config) => {
             globalThis.controlCounter++
         }
 
-        if (OK_MAX) globalThis.restartTimerMax = 0
-        if (OK_UNV) globalThis.restartTimerUnv = 0
         if (OK_PREV) globalThis.restartTimerPrev = 0
         if (OK_SYNCED) globalThis.restartTimerNotSynced = 0